Tests/LibWeb/Text/expected/css/computed-style-map-counters.txt

reset: chapter 2 section 5
increment: item 3
set: item 7
initial: none
